+++ src/main/java/net/thedevilthing/firstmod/Firstmod.java
@@ -18,6 +18,7 @@ import net.thedevilthing.firstmod.block.ModBlocks;
 import net.thedevilthing.firstmod.component.ModDataComponents;
 import net.thedevilthing.firstmod.item.ModCreativeModeTabs;
 import net.thedevilthing.firstmod.item.ModItems;
+import net.thedevilthing.firstmod.loot.ModLootConditions;
 import net.thedevilthing.firstmod.loot.ModLootModifiers;
 import net.thedevilthing.firstmod.util.datapack.DataPackGenerator;
 import net.thedevilthing.firstmod.util.datapack.DataPackManager;
@@ -54,6 +55,8 @@ public class Firstmod {
         ModDataComponents.register(modEventBus);
 
         ModLootModifiers.register(modEventBus);
+
+        ModLootConditions.register(modEventBus);
 
         // Register the item to a creative tab
         modEventBus.addListener(this::addCreative);