+++ src/main/java/net/thedevilthing/firstmod/datagen/ModDatapackProvider.java
@@ -7,6 +7,7 @@ import net.minecraft.data.PackOutput;
 import net.neoforged.neoforge.common.data.DatapackBuiltinEntriesProvider;
 import net.neoforged.neoforge.registries.NeoForgeRegistries;
 import net.thedevilthing.firstmod.Firstmod;
+import net.thedevilthing.firstmod.enchantment.ModEnchantments;
 import net.thedevilthing.firstmod.worldgen.ModBiomeModifiers;
 import net.thedevilthing.firstmod.worldgen.ModConfiguredFeatures;
 import net.thedevilthing.firstmod.worldgen.ModPlacedFeatures;
@@ -16,6 +17,7 @@ import java.util.concurrent.CompletableFuture;
 
 public class ModDatapackProvider extends DatapackBuiltinEntriesProvider {
     public static final RegistrySetBuilder BUILDER = new RegistrySetBuilder()
+            .add(Registries.ENCHANTMENT, ModEnchantments::bootstrap)
             .add(Registries.CONFIGURED_FEATURE, ModConfiguredFeatures::bootstrap)
             .add(Registries.PLACED_FEATURE, ModPlacedFeatures::bootstrap)
             .add(NeoForgeRegistries.Keys.BIOME_MODIFIERS, ModBiomeModifiers::bootstrap);